🎨 Move pf2e.Pf2eStat to pf2e.qute.QuteGenericStat
